% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.chr}
\alias{recode.chr}
\title{Recode Character or factor Variables
重新编码字符或因子变量}
\usage{
\code{recode.chr(df$var)}
}
\arguments{
\item{x}{A character or factor vector to be recoded.}
}
\value{
Invisibly returns the input vector. Main purpose is side effect of
generating recode code in RStudio.

返回\code{case_when()}结构的重编码代码。
}
\description{
\code{recode.chr()} Directly identifies variable categories Insert
\code{case_when()} code in RStudio to help quickly recoding character variables.

\code{recode.chr()}直接识别变量类目在RStudio中插入\code{case_when()}代码，
帮助快速重新编码。
}
\examples{
# 创建示例数据
df <- data.frame(
  age = c(28, 20, 30, 35, 36, "> 80 year", "<20 year"),
  edu = c("college or above", "Below high school", "High school",
         "College or above", "Below high school", "High school",
         "College, or Above")
)

\dontrun{
recode.chr(df$edu)
}

# 生成的代码如下,

# recode.chr(df$edu)
df$edu <- case_when(
  df$edu == "college or above" ~ "college or above",
  df$edu == "Below high school" ~ "Below high school",
  df$edu == "High school" ~ "High school",
  df$edu == "College or above" ~ "College or above",
  df$edu == "College, or Above" ~ "College, or Above",
  TRUE ~ NA_character_
)

}
