% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.num}
\alias{recode.num}
\title{Recode non-numeric numeric variables
重新编码非数值型的数值变量}
\usage{
\code{recode.chr(df$var)}
}
\arguments{
\item{x}{A vector to be converted to numeric}
}
\value{
Invisibly returns the input vector. Main purpose is side effect of
generating recode code in RStudio.

返回\code{case_when()}结构的重编码代码。
}
\description{
\code{recode.chr()} Directly identifies non-numeric items Insert
\code{case_when()} code in RStudio to help quickly recoding numeric variables.

\code{recode.chr()}直接识别非数字类目在RStudio中插入\code{case_when()}代码，
帮助快速重新编码。
}
\examples{
# 创建示例数据
df <- data.frame(
  age = c(28, 20, 30, 35, 36, "> 80 year", "<20 year"),
  edu = c("college or above", "Below high school", "High school",
         "College or above", "Below high school", "High school",
         "College, or Above")
)

\dontrun{
recode.num(df$age)
}

# 生成的代码如下,

# recode.num(df$age)
df$age <- case_when(
  df$age == "> 80 year" ~ "> 80 year",
  df$age == "<20 year" ~ "<20 year",
  TRUE ~ df$age
) \%>\% as.numeric()

}
