% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyds_miss.R
\name{yyds_miss}
\alias{yyds_miss}
\title{计算变量缺失数量和缺失占比}
\usage{
yyds_miss(data)
}
\arguments{
\item{data}{需要分析缺失值的数据框或tibble}
}
\value{
返回包含三列的数据框：
\item{variable}{字符型。变量/列名}
\item{missing_count}{整型。变量中的缺失值数量（NA）}
\item{missing_pct}{数值型。缺失值百分比，保留1位小数}
}
\description{
该函数用于计算数据框中每个变量的缺失值（NA）数量及百分比，
返回包含变量名、缺失数量和缺失百分比的数据框。
}
\examples{
df <- data.frame(
  年龄 = c(25, NA, 30, 35, NA),
  性别 = c("男", "女", NA, "女", "男"),
  分数 = c(85, 90, NA, NA, 88)
)
yyds_miss(df)

}
